moved actor code exists check into Lpa service
